Move AppGifteeButton component to matters dir
