+++ pages/index.vue
@@ -81,7 +81,7 @@
 import links from '~/modules/links.json'
 import skills from '~/modules/skills.json'
 import AppBooks from '~/components/matters/AppBooks'
-import AppGifteeButton from '~/components/AppGifteeButton'
+import AppGifteeButton from '~/components/matters/AppGifteeButton'
 import AppHeader1 from '~/components/elements/AppHeader1'
 import AppHeader2 from '~/components/elements/AppHeader2'
 import AppLinkButton from '~/components/elements/AppLinkButton'
